Move all types into DataModels namespace
There are so many types for this, we really should keep it in a separate
namespace
